fix:刪除不必要檔案
